+++ src/scheduler/Compare.py
@@ -65,7 +65,7 @@ class CompareScheduler:
 
     def _compare_single_run(self, compares_done):
         try:
-            compare_id, redo = self.in_queue.get(timeout=int(self.config['ExpertSettings']['block_delay']))
+            compare_id, redo = self.in_queue.get(timeout=float(self.config['ExpertSettings']['block_delay']))
         except Empty:
             pass
         else:
